+++ src/ezpadova/__init__.py
@@ -1,6 +1,7 @@
 from .deprecated import get_Z_isochrones, get_one_isochrone, get_t_isochrones
-# from . import parsec
+from .parsec import get_isochrones
+from . import parsec
 
 __version__ = "2.0.0"
 
-__all__ = ["get_Z_isochrones", "get_one_isochrone", "get_t_isochrones", "parsec"]
+__all__ = ["get_isochrones", "get_Z_isochrones", "get_one_isochrone", "get_t_isochrones", "parsec"]
